Add teams to fine
