+++ src/Server/Models/Fine.cs
@@ -4,11 +4,12 @@ namespace DynamoLeagueBlazor.Server.Models;
 
 public record Fine : BaseEntity
 {
-    public Fine(decimal amount, string reason, int playerId)
+    public Fine(decimal amount, string reason, int playerId, int teamId)
     {
         Amount = amount;
         Reason = reason;
         PlayerId = playerId;
+        TeamId = teamId;
     }
 
     [Column(TypeName = "decimal(18,0)")]
@@ -17,6 +18,8 @@ public record Fine : BaseEntity
     public DateTime CreatedOn { get; set; } = DateTime.Now;
     public string Reason { get; set; }
     public int PlayerId { get; set; }
+    public int TeamId { get; set; }
 
     public Player Player { get; private set; } = null!;
+    public Team Team { get; private set; } = null!;
 }
