+++ src/Tests/Features/Dashboard/TopOffendersTests.cs
@@ -21,11 +21,15 @@ public class TopOffendersTests : IntegrationTestBase
     public async Task GivenAnyAuthenticatedUser_WhenThereIsOnePlayerWithAFine_ThenReturnsOnePlayerWithAFine()
     {
         var application = CreateUserAuthenticatedApplication();
+
+        var stubTeam = CreateFakeTeam();
+        await application.AddAsync(stubTeam);
+
         var mockPlayer = CreateFakePlayer();
-        await application.AddAsync(mockPlayer);
-        var mockFine = CreateFakeFine(mockPlayer.Id);
+        mockPlayer.TeamId = stubTeam.Id;
+        var mockFine = mockPlayer.AddFine(int.MaxValue, RandomString);
         mockFine.Status = true;
-        await application.AddAsync(mockFine);
+        await application.AddAsync(mockPlayer);
 
         var client = application.CreateClient();
 
@@ -44,24 +48,24 @@ public class TopOffendersTests : IntegrationTestBase
     public async Task GivenAnyAuthenticatedUser_WhenThereIsElevenPlayersWithApprovedFines_ThenReturnsOnlyTopTenByFineAmount()
     {
         var application = CreateUserAuthenticatedApplication();
+
+        var stubTeam = CreateFakeTeam();
+        await application.AddAsync(stubTeam);
+
         foreach (var count in Enumerable.Range(0, 10))
         {
             var mockPlayer = CreateFakePlayer();
+            mockPlayer.TeamId = stubTeam.Id;
+            var fine = mockPlayer.AddFine(int.MaxValue, RandomString);
+            fine.Status = true;
             await application.AddAsync(mockPlayer);
-
-            var mockFine = CreateFakeFine(mockPlayer.Id);
-            mockFine.Status = true;
-            mockFine.Amount = int.MaxValue;
-            await application.AddAsync(mockFine);
         }
 
-        var sixthPlayer = CreateFakePlayer();
-        await application.AddAsync(sixthPlayer);
-
-        var lowestFine = CreateFakeFine(sixthPlayer.Id);
+        var eleventhPlayerWithFine = CreateFakePlayer();
+        eleventhPlayerWithFine.TeamId = stubTeam.Id;
+        var lowestFine = eleventhPlayerWithFine.AddFine(int.MinValue, RandomString);
         lowestFine.Status = true;
-        lowestFine.Amount = 1;
-        await application.AddAsync(lowestFine);
+        await application.AddAsync(eleventhPlayerWithFine);
 
         var client = application.CreateClient();
